pyproject.toml
[project]
name = "python-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.11.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
